renamed #nominal_level #level:
When I started developing this parser, I thought I had to distinguish between real (or actual) and nominal levels
(for handling mal-formed lists for example), but so far such distinction has not been reflected in the current
implementation. So I will use the shorter name for now and rethink about the distinction when the need arises.